feat: implement SQLAlchemy models for users, cafeteria management, products, sales, and waste tracking

+++ backend/app/models/sales.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Float, DateTime, ForeignKey
+from sqlalchemy import Column, String, Float, DateTime, ForeignKey, Integer
 from sqlalchemy.sql import func
 from app.core.database import Base
 
@@ -8,7 +8,7 @@ class Sale(Base):
     id = Column(String, primary_key=True, index=True)
     user_id = Column(String, index=True)
     product_id = Column(String, ForeignKey("products.id"))
-    quantity = Column(Float)
+    quantity = Column(Integer)
     total_price = Column(Float)
     timestamp = Column(DateTime(timezone=True), server_default=func.now())
     payment_method = Column(String)
@@ -18,6 +18,6 @@ class Waste(Base):
 
     id = Column(String, primary_key=True, index=True)
     product_id = Column(String, ForeignKey("products.id"))
-    quantity = Column(Float)
+    quantity = Column(Integer)
     reason = Column(String)
     timestamp = Column(DateTime(timezone=True), server_default=func.now())
